move some files and refactor code
